add perform action and disconnect endpoints, restructure files
